move json2 to dot/rpc
